+++ test/i18n_alchemy/numeric_parser_test.rb
@@ -44,8 +44,13 @@ class NumericParserTest < MiniTest::Unit::TestCase
   end
 
   def test_localize_numeric_values_with_delimiters
-    assert_equal "1001.20", @parser.localize(1001.2)
-    assert_equal "1000001.20", @parser.localize(1_000_001.2)
+    assert_equal "123.20", @parser.localize(123.2)
+    assert_equal "999.27", @parser.localize(999.27)
+  end
+
+  def test_localize_numeric_values_with_thousand_separators
+    assert_equal "1,001.20", @parser.localize(1001.2)
+    assert_equal "1,000,001.20", @parser.localize(1_000_001.2)
   end
 
   def test_localize_numbers_based_on_current_i18n_locale
@@ -56,7 +61,13 @@ class NumericParserTest < MiniTest::Unit::TestCase
 
   def test_localize_numbers_with_delimiters_based_on_current_i18n_locale
     I18n.with_locale :pt do
-      assert_equal "1001,20", @parser.localize(1_001.2)
+      assert_equal "99,20", @parser.localize(99.2)
+    end
+  end
+
+  def test_localize_numbers_with_separators_based_on_current_i18n_locale
+    I18n.with_locale :pt do
+      assert_equal "1.001,20", @parser.localize(1_001.2)
     end
   end
 end