+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 * Localize attributes=, assign_attributes, update_attribute, and update_attributes ([@tomas-stefano](https://github.com/tomas-stefano))
 * Localize nested attributes ([@tomas-stefano](https://github.com/tomas-stefano))
 * JRuby compatibility ([@sobrinho](https://github.com/sobrinho))
-* Format numeric values based on the current I18n number precision
+* Format numeric values based on the current I18n number precision, delimiter and thousand separator